Split out ambient music so it's not client only
